@@ -95,49 +86,71 @@
     "babel-preset-stage-2": "^6.24.1",
     "babel-register": "^6.26.0",
     "codecov.io": "^0.1.6",
-    "cross-env": "^5.1.3",
+    "cross-env": "^5.1.6",
     "deep-equal": "^1.0.1",
-    "eslint": "^4.18.1",
+    "eslint": "^4.19.1",
     "eslint-config-react-app": "^2.1.0",
-    "eslint-plugin-babel": "^4.1.2",
-    "eslint-plugin-flowtype": "^2.46.1",
-    "eslint-plugin-import": "^2.9.0",
+    "eslint-plugin-babel": "^5.1.0",
+    "eslint-plugin-flowtype": "^2.48.0",
+    "eslint-plugin-import": "^2.12.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0",
+    "eslint-plugin-react": "^7.8.2",
     "flow-bin": "^0.54.1",
-    "flux-standard-action": "^2.0.1",
+    "flux-standard-action": "^2.0.3",
+    "glow": "^1.2.2",
     "husky": "^0.14.3",
     "immutable": "^3.8.2",
-    "jest": "^22.4.2",
-    "lint-staged": "^7.0.0",
-    "lodash-webpack-plugin": "^0.11.2",
-    "prettier": "^1.11.1",
+    "jest": "^23.1.0",
+    "lint-staged": "^7.1.2",
+    "lodash-webpack-plugin": "^0.11.5",
+    "prettier": "^1.13.3",
     "prettier-eslint-cli": "^4.7.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
     "react-redux": "^5.0.7",
-    "redux": "^3.7.2",
+    "redux": "^4.0.0",
     "redux-immutablejs": "^0.0.8",
     "rifraf": "^2.0.3",
     "rimraf": "^2.5.4",
-    "stringstream": "^0.0.5",
+    "stringstream": "0.0.6",
     "tmp": "0.0.33",
-    "webpack": "^3.11.0",
-    "webpack-bundle-analyzer": "^2.11.1"
+    "webpack": "^4.10.2",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-cli": "^2.1.4"
   },
   "peerDependencies": {
     "react": "^15.0.0-0 || ^16.0.0-0",
     "react-redux": "^4.3.0 || ^5.0.0",
     "redux": "^3.0.0 || ^4.0.0"
   },
-  "files": ["README.md", "es", "lib", "dist", "immutable.js", "*.js.flow"],
+  "files": [
+    "README.md",
+    "es",
+    "lib",
+    "dist",
+    "immutable.js",
+    "*.js.flow"
+  ],
   "lint-staged": {
-    "*.{js,json,md,css}": ["prettier --write", "git add"]
+    "*.{js,json,md,css}": [
+      "prettier --write",
+      "git add"
+    ]
   },
   "jest": {
-    "collectCoverageFrom": ["src/**/*.js", "!src/**/__tests__/**/*.js"],
-    "coverageReporters": ["text", "lcov", "html"],
-    "coveragePathIgnorePatterns": ["/node_modules/", "/dist/"],
+    "collectCoverageFrom": [
+      "src/**/*.js",
+      "!src/**/__tests__/**/*.js"
+    ],
+    "coverageReporters": [
+      "text",
+      "lcov",
+      "html"
+    ],
+    "coveragePathIgnorePatterns": [
+      "/node_modules/",
+      "/dist/"
+    ],
     "testRegex": "__tests__/.*\\.spec\\.js$",
     "testEnvironment": "jsdom"
   },
