@@ -37,7 +37,7 @@ const createField = (structure: Structure<*, *>) => {
       return shallowCompare(this, nextProps, nextState)
     }
 
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
       this.context._reduxForm.register(
         this.name,
         'Field',
@@ -46,7 +46,7 @@ const createField = (structure: Structure<*, *>) => {
       )
     }
 
-    componentWillReceiveProps(nextProps: Props, nextContext: any) {
+    UNSAFE_componentWillReceiveProps(nextProps: Props, nextContext: any) {
       const oldName = prefixName(this.context, this.props.name)
       const newName = prefixName(nextContext, nextProps.name)
 
