@@ -1,4 +1,5 @@
 // @flow
+import * as React from 'react'
 import hoistStatics from 'hoist-non-react-statics'
 import invariant from 'invariant'
 import isPromise from 'is-promise'
@@ -212,6 +213,7 @@ export type Props = {
   asyncValidating: boolean,
   blur: BlurAction,
   change: ChangeAction,
+  children?: React.Node,
   clearSubmit: ClearSubmitAction,
   destroy: DestroyAction,
   destroyOnUnmount: boolean,
@@ -236,6 +238,7 @@ export type Props = {
   onSubmitSuccess?: OnSubmitSuccess,
   pristine: boolean,
   propNamespace?: string,
+  pure?: boolean,
   registeredFields: Array<{ name: string, type: FieldType, count: number }>,
   registerField: RegisterFieldAction,
   reset: ResetAction,
@@ -302,6 +305,7 @@ const createReduxForm = (structure: Structure<*, *>) => {
         static WrappedComponent: ComponentType<*>
 
         context: ReactContext
+        wrapped: ?React.Component<*, *>
 
         destroyed = false
         fieldCounts = {}
@@ -513,7 +517,7 @@ const createReduxForm = (structure: Structure<*, *>) => {
           }
         }
 
-        componentWillMount() {
+        UNSAFE_componentWillMount() {
           if (!isHotReloading()) {
             this.initIfNeeded()
             this.validateIfNeeded()
@@ -825,6 +829,10 @@ const createReduxForm = (structure: Structure<*, *>) => {
 
         reset = (): void => this.props.reset()
 
+        saveRef = (ref: ?React.Component<*, *>) => {
+          this.wrapped = ref
+        }
+
         render() {
           // remove some redux-form config-only props
           /* eslint-disable no-unused-vars */
@@ -935,7 +943,7 @@ const createReduxForm = (structure: Structure<*, *>) => {
             ...rest
           }
           if (isClassComponent(WrappedComponent)) {
-            propsToPass.ref = 'wrapped'
+            ;((propsToPass: any): Object).ref = this.saveRef
           }
           return createElement(WrappedComponent, propsToPass)
         }
@@ -1128,14 +1136,16 @@ const createReduxForm = (structure: Structure<*, *>) => {
 
         get wrappedInstance(): ?Component<*, *> {
           // for testing
-          return this.ref && this.ref.getWrappedInstance().refs.wrapped
+          return this.ref && this.ref.getWrappedInstance().wrapped
         }
 
         render() {
           const { initialValues, ...rest } = this.props
           return createElement(ConnectedForm, {
             ...rest,
-            ref: (ref: ?ConnectedComponent<Form>) => (this.ref = ref),
+            ref: (ref: ?ConnectedComponent<Form>) => {
+              this.ref = ref
+            },
             // convert initialValues if need to
             initialValues: fromJS(initialValues)
           })
