@@ -1,5 +1,5 @@
 // @flow
-import { Component, createElement } from 'react'
+import * as React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import createFieldProps from './createFieldProps'
@@ -28,10 +28,11 @@ const createConnectedFields = (structure: Structure<*, *>) => {
     return warning && warning._warning ? warning._warning : warning
   }
 
-  class ConnectedFields extends Component<Props> {
+  class ConnectedFields extends React.Component<Props> {
     onChangeFns = {}
     onFocusFns = {}
     onBlurFns = {}
+    ref: ?React.Component<*>
 
     constructor(props: Props) {
       super(props)
@@ -89,7 +90,7 @@ const createConnectedFields = (structure: Structure<*, *>) => {
     }
 
     getRenderedComponent() {
-      return this.refs.renderedComponent
+      return this.ref
     }
 
     handleChange = (name: string, event: any): void => {
@@ -122,6 +123,10 @@ const createConnectedFields = (structure: Structure<*, *>) => {
       }
     }
 
+    saveRef = (ref: ?React.Component<*>) => {
+      this.ref = ref
+    }
+
     render() {
       const { component, withRef, _fields, _reduxForm, ...rest } = this.props
       const { sectionPrefix, form } = _reduxForm
@@ -145,10 +150,10 @@ const createConnectedFields = (structure: Structure<*, *>) => {
         {}
       )
       if (withRef) {
-        props.ref = 'renderedComponent'
+        props.ref = this.saveRef
       }
 
-      return createElement(component, { ...props, ...custom })
+      return React.createElement(component, { ...props, ...custom })
     }
   }
 
