@@ -1,4 +1,4 @@
-import createCreateDeleteInWithCleanUp from '../deleteInWithCleanUp'
+import createDeleteInWithCleanUp from '../deleteInWithCleanUp'
 import plain from '../structure/plain'
 import plainExpectations from '../structure/plain/__tests__/expectations'
 import immutable from '../structure/immutable'
@@ -6,7 +6,7 @@ import immutableExpectations from '../structure/immutable/__tests__/expectations
 
 const describeDeleteInWithCleanUp = (name, structure, setup) => {
   const { fromJS } = structure
-  const deleteInWithCleanUp = createCreateDeleteInWithCleanUp(structure)()
+  const deleteInWithCleanUp = createDeleteInWithCleanUp(structure)()
 
   describe(name, () => {
     beforeAll(() => {
@@ -147,8 +147,8 @@ const describeDeleteInWithCleanUp = (name, structure, setup) => {
     })
 
     it('should only delete cats because I am a dog person', () => {
-      const validation = (structure) => (state, path) => path.startsWith('cat')
-      const deleteInSpecial = createCreateDeleteInWithCleanUp(structure)(validation)
+      const validation = structure => (state, path) => path.startsWith('cat')
+      const deleteInSpecial = createDeleteInWithCleanUp(structure)(validation)
 
       expect(
         deleteInSpecial(
@@ -171,7 +171,7 @@ const describeDeleteInWithCleanUp = (name, structure, setup) => {
           'cat'
         )
       ).toEqualMap({
-        dog: 'Scooby',
+        dog: 'Scooby'
       })
     })
   })
