@@ -37,7 +37,7 @@ import {
   CLEAR_FIELDS,
   UPDATE_SYNC_WARNINGS
 } from './actionTypes'
-import createCreateDeleteInWithCleanUp from './deleteInWithCleanUp'
+import createDeleteInWithCleanUp from './deleteInWithCleanUp'
 import plain from './structure/plain'
 import type { Action, Structure } from './types.js.flow'
 
@@ -48,9 +48,11 @@ const shouldDelete = ({ getIn }) => (state, path) => {
     initialValuesPath = path.replace('values', 'initial')
   }
 
-  const initialValueComparison = initialValuesPath ? (getIn(state, initialValuesPath) === undefined) : true
+  const initialValueComparison = initialValuesPath
+    ? getIn(state, initialValuesPath) === undefined
+    : true
 
-  return (getIn(state, path) !== undefined) && initialValueComparison
+  return getIn(state, path) !== undefined && initialValueComparison
 }
 
 const isReduxFormAction = action =>
@@ -73,8 +75,10 @@ function createReducer<M, L>(structure: Structure<M, L>) {
     some,
     splice
   } = structure
-  const deleteInWithCleanUp = createCreateDeleteInWithCleanUp(structure)(shouldDelete)
-  const plainDeleteInWithCleanUp = createCreateDeleteInWithCleanUp(plain)(shouldDelete)
+  const deleteInWithCleanUp = createDeleteInWithCleanUp(structure)(shouldDelete)
+  const plainDeleteInWithCleanUp = createDeleteInWithCleanUp(plain)(
+    shouldDelete
+  )
   const doSplice = (state, key, field, index, removeNum, value, force) => {
     const existing = getIn(state, `${key}.${field}`)
     return existing || force
@@ -135,10 +139,21 @@ function createReducer<M, L>(structure: Structure<M, L>) {
   }
 
   const behaviors: { [string]: { (state: any, action: Action): M } } = {
-    [ARRAY_INSERT](state, { meta: { field, index }, payload }) {
+    [ARRAY_INSERT](
+      state,
+      {
+        meta: { field, index },
+        payload
+      }
+    ) {
       return arraySplice(state, field, index, 0, payload)
     },
-    [ARRAY_MOVE](state, { meta: { field, from, to } }) {
+    [ARRAY_MOVE](
+      state,
+      {
+        meta: { field, from, to }
+      }
+    ) {
       const array = getIn(state, `values.${field}`)
       const length = array ? size(array) : 0
       let result = state
@@ -158,31 +173,68 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       }
       return result
     },
-    [ARRAY_POP](state, { meta: { field } }) {
+    [ARRAY_POP](
+      state,
+      {
+        meta: { field }
+      }
+    ) {
       const array = getIn(state, `values.${field}`)
       const length = array ? size(array) : 0
       return length ? arraySplice(state, field, length - 1, 1) : state
     },
-    [ARRAY_PUSH](state, { meta: { field }, payload }) {
+    [ARRAY_PUSH](
+      state,
+      {
+        meta: { field },
+        payload
+      }
+    ) {
       const array = getIn(state, `values.${field}`)
       const length = array ? size(array) : 0
       return arraySplice(state, field, length, 0, payload)
     },
-    [ARRAY_REMOVE](state, { meta: { field, index } }) {
+    [ARRAY_REMOVE](
+      state,
+      {
+        meta: { field, index }
+      }
+    ) {
       return arraySplice(state, field, index, 1)
     },
-    [ARRAY_REMOVE_ALL](state, { meta: { field } }) {
+    [ARRAY_REMOVE_ALL](
+      state,
+      {
+        meta: { field }
+      }
+    ) {
       const array = getIn(state, `values.${field}`)
       const length = array ? size(array) : 0
       return length ? arraySplice(state, field, 0, length) : state
     },
-    [ARRAY_SHIFT](state, { meta: { field } }) {
+    [ARRAY_SHIFT](
+      state,
+      {
+        meta: { field }
+      }
+    ) {
       return arraySplice(state, field, 0, 1)
     },
-    [ARRAY_SPLICE](state, { meta: { field, index, removeNum }, payload }) {
+    [ARRAY_SPLICE](
+      state,
+      {
+        meta: { field, index, removeNum },
+        payload
+      }
+    ) {
       return arraySplice(state, field, index, removeNum, payload)
     },
-    [ARRAY_SWAP](state, { meta: { field, indexA, indexB } }) {
+    [ARRAY_SWAP](
+      state,
+      {
+        meta: { field, indexA, indexB }
+      }
+    ) {
       let result = state
       rootKeys.forEach(key => {
         const valueA = getIn(result, `${key}.${field}[${indexA}]`)
@@ -194,10 +246,22 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       })
       return result
     },
-    [ARRAY_UNSHIFT](state, { meta: { field }, payload }) {
+    [ARRAY_UNSHIFT](
+      state,
+      {
+        meta: { field },
+        payload
+      }
+    ) {
       return arraySplice(state, field, 0, 0, payload)
     },
-    [AUTOFILL](state, { meta: { field }, payload }) {
+    [AUTOFILL](
+      state,
+      {
+        meta: { field },
+        payload
+      }
+    ) {
       let result: any = state
       result = deleteInWithCleanUp(result, `asyncErrors.${field}`)
       result = deleteInWithCleanUp(result, `submitErrors.${field}`)
@@ -205,7 +269,13 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       result = setIn(result, `values.${field}`, payload)
       return result
     },
-    [BLUR](state, { meta: { field, touch }, payload }) {
+    [BLUR](
+      state,
+      {
+        meta: { field, touch },
+        payload
+      }
+    ) {
       let result = state
       const initial = getIn(result, `initial.${field}`)
       if (initial === undefined && payload === '') {
@@ -225,7 +295,10 @@ function createReducer<M, L>(structure: Structure<M, L>) {
     },
     [CHANGE](
       state,
-      { meta: { field, touch, persistentSubmitErrors }, payload }
+      {
+        meta: { field, touch, persistentSubmitErrors },
+        payload
+      }
     ) {
       let result = state
       const initial = getIn(result, `initial.${field}`)
@@ -254,12 +327,19 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       result = deleteIn(result, 'error')
       return result
     },
-    [CLEAR_ASYNC_ERROR](state, { meta: { field } }) {
+    [CLEAR_ASYNC_ERROR](
+      state,
+      {
+        meta: { field }
+      }
+    ) {
       return deleteIn(state, `asyncErrors.${field}`)
     },
     [CLEAR_FIELDS](
       state,
-      { meta: { keepTouched, persistentSubmitErrors, fields } }
+      {
+        meta: { keepTouched, persistentSubmitErrors, fields }
+      }
     ) {
       let result = state
       fields.forEach(field => {
@@ -281,7 +361,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
         : deleteIn(result, 'anyTouched')
       return result
     },
-    [FOCUS](state, { meta: { field } }) {
+    [FOCUS](
+      state,
+      {
+        meta: { field }
+      }
+    ) {
       let result = state
       const previouslyActive = getIn(state, 'active')
       result = deleteIn(result, `fields.${previouslyActive}.active`)
@@ -412,7 +497,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       result = setIn(result, 'initial', newInitialValues)
       return result
     },
-    [REGISTER_FIELD](state, { payload: { name, type } }) {
+    [REGISTER_FIELD](
+      state,
+      {
+        payload: { name, type }
+      }
+    ) {
       const key = `registeredFields['${name}']`
       let field = getIn(state, key)
       if (field) {
@@ -436,7 +526,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       }
       return result
     },
-    [RESET_SECTION](state, { meta: { sections } }) {
+    [RESET_SECTION](
+      state,
+      {
+        meta: { sections }
+      }
+    ) {
       let result = state
 
       sections.forEach(section => {
@@ -462,7 +557,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
     [SUBMIT](state) {
       return setIn(state, 'triggerSubmit', true)
     },
-    [START_ASYNC_VALIDATION](state, { meta: { field } }) {
+    [START_ASYNC_VALIDATION](
+      state,
+      {
+        meta: { field }
+      }
+    ) {
       return setIn(state, 'asyncValidating', field || true)
     },
     [START_SUBMIT](state) {
@@ -509,7 +609,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       }
       return result
     },
-    [SET_SUBMIT_FAILED](state, { meta: { fields } }) {
+    [SET_SUBMIT_FAILED](
+      state,
+      {
+        meta: { fields }
+      }
+    ) {
       let result = state
       result = setIn(result, 'submitFailed', true)
       result = deleteIn(result, 'submitSucceeded')
@@ -528,7 +633,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       result = setIn(result, 'submitSucceeded', true)
       return result
     },
-    [TOUCH](state, { meta: { fields } }) {
+    [TOUCH](
+      state,
+      {
+        meta: { fields }
+      }
+    ) {
       let result = state
       fields.forEach(
         field => (result = setIn(result, `fields.${field}.touched`, true))
@@ -536,7 +646,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       result = setIn(result, 'anyTouched', true)
       return result
     },
-    [UNREGISTER_FIELD](state, { payload: { name, destroyOnUnmount } }) {
+    [UNREGISTER_FIELD](
+      state,
+      {
+        payload: { name, destroyOnUnmount }
+      }
+    ) {
       let result = state
       const key = `registeredFields['${name}']`
       let field = getIn(result, key)
@@ -577,7 +692,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       }
       return result
     },
-    [UNTOUCH](state, { meta: { fields } }) {
+    [UNTOUCH](
+      state,
+      {
+        meta: { fields }
+      }
+    ) {
       let result = state
       fields.forEach(
         field => (result = deleteIn(result, `fields.${field}.touched`))
@@ -590,7 +710,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
         : deleteIn(result, 'anyTouched')
       return result
     },
-    [UPDATE_SYNC_ERRORS](state, { payload: { syncErrors, error } }) {
+    [UPDATE_SYNC_ERRORS](
+      state,
+      {
+        payload: { syncErrors, error }
+      }
+    ) {
       let result = state
       if (error) {
         result = setIn(result, 'error', error)
@@ -606,7 +731,12 @@ function createReducer<M, L>(structure: Structure<M, L>) {
       }
       return result
     },
-    [UPDATE_SYNC_WARNINGS](state, { payload: { syncWarnings, warning } }) {
+    [UPDATE_SYNC_WARNINGS](
+      state,
+      {
+        payload: { syncWarnings, warning }
+      }
+    ) {
       let result = state
       if (warning) {
         result = setIn(result, 'warning', warning)
